requirements.txt
absl-py==2.1.0
alembic==1.13.2
astunparse==1.6.3
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
filelock==3.16.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
googlemaps==4.10.0
greenlet==3.1.0
grpcio==1.66.1
h5py==3.11.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
overpy==0.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.2
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
rich==13.8.1
scipy==1.14.1
seaborn==0.13.2
setuptools==74.1.2
six==1.16.0
SQLAlchemy==2.0.34
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.94
ultralytics-thop==2.0.6
urllib3==2.2.3
virtualenv==20.26.4
visitor==0.1.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
WTForms==3.1.2
